:hammer: Unificación metodos
