+++ routes/routes.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
 const joyasRoutes = require("./joyasRoutes");
-const routesQueryString = require("./routesQueryString");
 
-router.use("/joyas", joyasRoutes, routesQueryString);
+
+router.use("/joyas", joyasRoutes);
 router.get("*", (req, res) => {
     res.status(404).json({ message: "Ruta no encontrada" });
 });
